+++ link/link-graph-service-with-loco/src/models/_entities/prelude.rs
@@ -3,3 +3,4 @@
 pub use super::consumer_offsets::Entity as ConsumerOffsets;
 pub use super::edges::Entity as Edges;
 pub use super::entity_presence::Entity as EntityPresence;
+pub use super::processed_events::Entity as ProcessedEvents;